change deployer to wallet
