--- transactions/p-chain/txs/helper.go
@@ -1,63 +1,0 @@
-// Copyright (C) 2025, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-package txs
-
-import (
-	"errors"
-	"fmt"
-	"time"
-
-	"github.com/ava-labs/avalanche-tooling-sdk-go/multisig"
-	utilsSDK "github.com/ava-labs/avalanche-tooling-sdk-go/utils"
-	"github.com/ava-labs/avalanche-tooling-sdk-go/wallet"
-	"github.com/ava-labs/avalanchego/ids"
-	commonAvago "github.com/ava-labs/avalanchego/wallet/subnet/primary/common"
-)
-
-func (d *PublicDeployer) Commit(ms multisig.Multisig, wallet wallet.Wallet, waitForTxAcceptance bool) (ids.ID, error) {
-	if ms.Undefined() {
-		return ids.Empty, multisig.ErrUndefinedTx
-	}
-	isReady, err := ms.IsReadyToCommit()
-	if err != nil {
-		return ids.Empty, err
-	}
-	if !isReady {
-		return ids.Empty, errors.New("tx is not fully signed so can't be committed")
-	}
-	tx, err := ms.GetWrappedPChainTx()
-	if err != nil {
-		return ids.Empty, err
-	}
-	const (
-		repeats             = 3
-		sleepBetweenRepeats = 2 * time.Second
-	)
-	var issueTxErr error
-	if err != nil {
-		return ids.Empty, err
-	}
-	for i := 0; i < repeats; i++ {
-		ctx, cancel := utilsSDK.GetAPILargeContext()
-		defer cancel()
-		options := []commonAvago.Option{commonAvago.WithContext(ctx)}
-		if !waitForTxAcceptance {
-			options = append(options, commonAvago.WithAssumeDecided())
-		}
-		// TODO: split error checking and recovery between issuing and waiting for status
-		issueTxErr = wallet.P().IssueTx(tx, options...)
-		if issueTxErr == nil {
-			break
-		}
-		if ctx.Err() != nil {
-			issueTxErr = fmt.Errorf("timeout issuing/verifying tx with ID %s: %w", tx.ID(), issueTxErr)
-		} else {
-			issueTxErr = fmt.Errorf("error issuing tx with ID %s: %w", tx.ID(), issueTxErr)
-		}
-		time.Sleep(sleepBetweenRepeats)
-	}
-	if issueTxErr != nil {
-		return ids.Empty, fmt.Errorf("issue tx error %w", issueTxErr)
-	}
-	return tx.ID(), issueTxErr
-}
