+++ wallet/wallet.go
@@ -4,14 +4,19 @@ package wallet
 
 import (
 	"context"
-
+	"errors"
+	"fmt"
 	"github.com/ava-labs/avalanche-tooling-sdk-go/keychain"
+	"github.com/ava-labs/avalanche-tooling-sdk-go/multisig"
+	utilsSDK "github.com/ava-labs/avalanche-tooling-sdk-go/utils"
 	"github.com/ava-labs/avalanchego/ids"
 	avagokeychain "github.com/ava-labs/avalanchego/utils/crypto/keychain"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 	"github.com/ava-labs/avalanchego/wallet/subnet/primary"
 	"github.com/ava-labs/avalanchego/wallet/subnet/primary/common"
+	commonAvago "github.com/ava-labs/avalanchego/wallet/subnet/primary/common"
+	"time"
 )
 
 type Wallet struct {
@@ -70,4 +75,70 @@ func (w *Wallet) SetSubnetAuthMultisig(authKeys []ids.ShortID) {
 
 func (w *Wallet) Addresses() []ids.ShortID {
 	return w.Keychain.Addresses().List()
+}
+
+func (w *Wallet) GetMultisigTxOptions(subnetAuthKeys []ids.ShortID) []common.Option {
+	options := []common.Option{}
+	walletAddrs := w.Keychain.Addresses().List()
+	changeAddr := walletAddrs[0]
+	// addrs to use for signing
+	customAddrsSet := set.Set[ids.ShortID]{}
+	customAddrsSet.Add(walletAddrs...)
+	customAddrsSet.Add(subnetAuthKeys...)
+	options = append(options, common.WithCustomAddresses(customAddrsSet))
+	// set change to go to wallet addr (instead of any other subnet auth key)
+	changeOwner := &secp256k1fx.OutputOwners{
+		Threshold: 1,
+		Addrs:     []ids.ShortID{changeAddr},
+	}
+	options = append(options, common.WithChangeOwner(changeOwner))
+	return options
+}
+
+func (w *Wallet) Commit(ms multisig.Multisig, waitForTxAcceptance bool) (ids.ID, error) {
+	if ms.Undefined() {
+		return ids.Empty, multisig.ErrUndefinedTx
+	}
+	isReady, err := ms.IsReadyToCommit()
+	if err != nil {
+		return ids.Empty, err
+	}
+	if !isReady {
+		return ids.Empty, errors.New("tx is not fully signed so can't be committed")
+	}
+	tx, err := ms.GetWrappedPChainTx()
+	if err != nil {
+		return ids.Empty, err
+	}
+	const (
+		repeats             = 3
+		sleepBetweenRepeats = 2 * time.Second
+	)
+	var issueTxErr error
+	if err != nil {
+		return ids.Empty, err
+	}
+	for i := 0; i < repeats; i++ {
+		ctx, cancel := utilsSDK.GetAPILargeContext()
+		defer cancel()
+		options := []commonAvago.Option{commonAvago.WithContext(ctx)}
+		if !waitForTxAcceptance {
+			options = append(options, commonAvago.WithAssumeDecided())
+		}
+		// TODO: split error checking and recovery between issuing and waiting for status
+		issueTxErr = w.P().IssueTx(tx, options...)
+		if issueTxErr == nil {
+			break
+		}
+		if ctx.Err() != nil {
+			issueTxErr = fmt.Errorf("timeout issuing/verifying tx with ID %s: %w", tx.ID(), issueTxErr)
+		} else {
+			issueTxErr = fmt.Errorf("error issuing tx with ID %s: %w", tx.ID(), issueTxErr)
+		}
+		time.Sleep(sleepBetweenRepeats)
+	}
+	if issueTxErr != nil {
+		return ids.Empty, fmt.Errorf("issue tx error %w", issueTxErr)
+	}
+	return tx.ID(), issueTxErr
 }
